tests/mock_panorama/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6

